+++ services/web/server/src/simcore_service_webserver/application.py
@@ -38,7 +38,7 @@ from .scicrunch.plugin import setup_scicrunch
 from .security import setup_security
 from .session import setup_session
 from .socketio.plugin import setup_socketio
-from .statics import setup_statics
+from .statics.plugin import setup_statics
 from .storage import setup_storage
 from .studies_dispatcher.plugin import setup_studies_dispatcher
 from .tags import setup_tags
