+++ services/web/server/src/simcore_service_webserver/statics/_constants.py
