+++ services/web/server/tests/unit/isolated/test_products_middlewares.py
@@ -13,7 +13,7 @@ from simcore_service_webserver.products._events import _set_app_state
 from simcore_service_webserver.products._middlewares import discover_product_middleware
 from simcore_service_webserver.products._model import Product
 from simcore_service_webserver.products.plugin import get_product_name
-from simcore_service_webserver.statics_constants import FRONTEND_APP_DEFAULT
+from simcore_service_webserver.statics._constants import FRONTEND_APP_DEFAULT
 from yarl import URL
